Address long lines, especially in JSON data
This commit makes some changes to long lines making them more readable. This especially affects JSON Test data, which are formatted one indent every collection level in order to make the display way more readable. While this leads to longer method lengths, it significantly improves on-sight understanding of the conents, which outweighs the disadvantages of long methods, which are already less important with regards to "data methods" unlike for "functionality methods"